cloud_inference/requirements/together_requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
datasets==3.5.0
dill==0.3.8
eval_type_backport==0.2.2
filelock==3.17.0
frozenlist==1.5.0
fsspec==2024.12.0
huggingface-hub==0.30.2
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
multiprocess==0.70.16
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.2.1
pyarrow==19.0.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
six==1.17.0
tabulate==0.9.0
together==1.4.0
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
xxhash==3.5.0
yarl==1.18.3